==> notebooks/05a_prediction_notebook.ipynb <==
LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"

TRACKING_URI = "file:///home/deena_gergis/iti/iti_e2e_live/notebooks/mlruns/"
EXPERIMENT_ID = "1"
RUN_ID = "4c35a9c9d26a48e8bfccfae05a63d348"

import os 
import sklearn
import pickle
import yaml

import pandas as pd

import mlflow
from mlflow.tracking import MlflowClient

## Load model 

artifact_path = os.path.join(TRACKING_URI.replace("file://", ""), 
                             EXPERIMENT_ID, 
                             RUN_ID, 
                             'artifacts')

# Load data pkl
data_path  = os.path.join(artifact_path, LOG_DATA_PKL)
with open(data_path, 'rb') as handle:
    data_pkl = pickle.load(handle)

# Load model pkl
model_path = os.path.join(artifact_path, LOG_MODEL_PKL)
with open(model_path, 'rb') as handle:
    model_pkl = pickle.load(handle)

model = model_pkl["model_object"]

model

## Predict sample entry 

CLUSTERS_YAML_PATH = "../data/processed/features_skills_clusters_description.yaml"

CLUSTERS_YAML_PATH

with open(CLUSTERS_YAML_PATH, "r") as stream:
    clusters_config = yaml.safe_load(stream)

molten_clusters = [(cluster_name, cluster_skill)
                   for cluster_name, cluster_skills in clusters_config.items()
                   for cluster_skill in cluster_skills]

clusters_df = pd.DataFrame(molten_clusters, columns=["cluster_name", "skill"])

### Recreate cluster features 

sample_skills = ['Pandas', 'TensorFlow', 'Torch/PyTorch', 'Python', 'Keras']

sample_clusters = clusters_df.copy()

sample_clusters["sample_skills"] = sample_clusters["skill"].isin(sample_skills)

cluster_features = sample_clusters.groupby("cluster_name")["sample_skills"].sum()

cluster_features

### Create OneHotEncoded skills 

features_names = pd.Series(data_pkl["features_names"])

skills_names = features_names[~features_names.isin(cluster_features.index)]

sample_skills

skills_names

skills_names = features_names[~features_names.isin(cluster_features.index)]

ohe_skills = pd.Series(skills_names.isin(sample_skills).astype(int).tolist(), 
                       index=skills_names)

ohe_skills

### Combine features 

features = pd.concat([ohe_skills, cluster_features])

features = features[data_pkl["features_names"]]

features

### Predict

predictions = model.predict_proba([features.values])
positive_probs = [prob[0][1] for prob in predictions]
pd.Series(positive_probs, 
          index=data_pkl["targets_names"]).sort_values(ascending=False)
