==> notebooks/05c_production_dummy_app.ipynb <==
PREDICTION_URL = 'http://127.0.0.1:5000/predict_jobs_probs'

import requests
import json
import pandas as pd

available_skills = ['Pandas', 'TensorFlow', 'Torch/PyTorch', 'Python']

response = requests.post(PREDICTION_URL, 
                         data=json.dumps(available_skills), 
                         headers={'content-type':'application/json'})

probs = json.loads(response.text)

pd.Series(probs).sort_values(ascending=False)
